refactor(profile): extract auth providers list to its own component and replace arrow for standard functions
