+++ src/components/dashboard/profile/update-profile-form.tsx
@@ -3,59 +3,12 @@
 import { ImagePreview } from "./image-preview";
 import { Button, Input, Label } from "@/components/ui";
 import { useToast } from "@/components/ui/use-toast";
-import { AUTH_PROVIDERS, AuthProvider } from "@/constants/auth";
+import { AuthProvider } from "@/constants/auth";
 import { useUpdateProfile } from "@/hooks/profile";
 import { useUploadThing } from "@/lib/files";
-import { cn } from "@/lib/utils";
 import { Loader } from "lucide-react";
 import { useState } from "react";
-
-type AuthProvidersListProps = {
-  userLinkedProviders: AuthProvider[];
-};
-
-function AuthProvidersList({ userLinkedProviders }: AuthProvidersListProps) {
-  return (
-    <section className="flex flex-col gap-2">
-      <header>
-        <h3 className="font-semibold text-center">Linked accounts</h3>
-      </header>
-
-      <ul className="flex justify-center gap-4">
-        {Object.keys(AUTH_PROVIDERS).map((provider) => {
-          // Iterates over the all the available providers and checks if the user has linked them
-          const isLinked = userLinkedProviders.includes(
-            provider as AuthProvider,
-          );
-          const { styles, Icon } = AUTH_PROVIDERS[provider as AuthProvider];
-
-          return (
-            <li key={provider}>
-              <Button
-                asChild
-                disabled={!isLinked}
-                size="icon"
-                className={cn(
-                  "p-2 pointer-events-none",
-                  styles.backgroundColor,
-                  !isLinked && "opacity-45",
-                )}
-              >
-                <Icon />
-              </Button>
-            </li>
-          );
-        })}
-      </ul>
-    </section>
-  );
-}
-
-type ProfileForm = {
-  name: string;
-  imageUrl: string;
-  file: File | null;
-};
+import { AuthProvidersList } from "./auth-providers-list";
 
 const FORM_IDS = {
   name: "name",
@@ -69,6 +22,12 @@ type UpdateProfileFormProps = {
   imageUrl: string;
 };
 
+type ProfileForm = {
+  name: string;
+  imageUrl: string;
+  file: File | null;
+};
+
 export function UpdateProfileForm({
   email,
   imageUrl,
@@ -79,30 +38,29 @@ export function UpdateProfileForm({
   const { mutateAsync, isPending: isUploadingForm } = useUpdateProfile();
   const { startUpload, isUploading: isUploadingImage } =
     useUploadThing("imageUploader");
-
-  const isLoading = isUploadingForm || isUploadingImage;
-
   const [profileForm, setProfileForm] = useState<ProfileForm>({
     name,
     imageUrl,
     file: null,
   });
 
-  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+  const isLoading = isUploadingForm || isUploadingImage;
+
+  function handleInputChange(event: React.ChangeEvent<HTMLInputElement>) {
     const { name, value } = event.target;
     setProfileForm((prev) => ({ ...prev, [name]: value }));
-  };
+  }
 
-  const handleImageChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+  function handleImageChange(event: React.ChangeEvent<HTMLInputElement>) {
     const file = event.target.files?.[0];
 
     if (file) {
       const imageUrl = URL.createObjectURL(file);
       setProfileForm((prev) => ({ ...prev, file, imageUrl }));
     }
-  };
+  }
 
-  const handleSubmitForm = async (event: React.FormEvent<HTMLFormElement>) => {
+  async function handleSubmitForm(event: React.FormEvent<HTMLFormElement>) {
     try {
       event.preventDefault();
 
@@ -134,7 +92,7 @@ export function UpdateProfileForm({
         variant: "destructive",
       });
     }
-  };
+  }
   return (
     <section className="w-full flex flex-col gap-4 items-center justify-center">
       <ImagePreview
